Separate server-side config struct into another file as well as agent-side
